Configure the tests to use VCR gem
Records an API calls and uses the recordings in subsequent test runs.
This allows the tests to be run without an api key and speeds them up
a little.

Configured to use the rspec :vcr tag to automatically use VCR in a test
and to filter out the OPEN_CAGE_API_KEY from recordings.